backend/requirements.txt
python-dotenv
qdrant-client
sentence-transformers
docling
PyMuPDF
langchain
langchain-experimental
tqdm
fastapi
uvicorn[standard]
google-generativeai
pydantic-settings
rank-bm25
httpx
neo4j